Up symbols cleanup from Component
